WIP: Add owner annotations to objects

+++ internal/synchronizer/synchronizer.go
@@ -218,8 +218,7 @@ func (s *Synchronizer[T, P]) Reconcile(ctx context.Context, req ctrl.Request) (c
 func (s *Synchronizer[T, P]) PerformActions(ctx context.Context, actions []action.Action) (ctrl.Result, error) {
 	var err error
 	for _, a := range actions {
-		// TODO: s.addOwnerAnnotation(a)
-		// Must handle IAMPolicyMember before adding owner annotation here
+		s.addOwnerAnnotation(a.GetObject(), a.GetOwner())
 		err = a.Do(ctx, s.client, s.scheme)
 		if err != nil {
 			return ctrl.Result{}, err
@@ -229,16 +228,73 @@ func (s *Synchronizer[T, P]) PerformActions(ctx context.Context, actions []actio
 	return ctrl.Result{}, nil
 }
 
-// TODO: Must handle IAMPolicyMember before adding owner annotation
-// func (s *Synchronizer[T, P]) addOwnerAnnotation(a action.Action) {
-// 	obj := a.GetObject()
-// 	annotations := obj.GetAnnotations()
-// 	if annotations == nil {
-// 		annotations = make(map[string]string)
-// 		obj.SetAnnotations(annotations)
-// 	}
-// 	annotations[s.ownerAnnotationKey] = client.ObjectKeyFromObject(a.GetOwner()).String()
-// }
+func (s *Synchronizer[T, P]) makeOwnerAnnotation(obj client.Object) string {
+	return client.ObjectKeyFromObject(obj).String()
+}
+
+func (s *Synchronizer[T, P]) HasOwnerAnnotation(obj, owner client.Object) bool {
+	ownerAnnotation := s.makeOwnerAnnotation(owner)
+	annotations := obj.GetAnnotations()
+	if annotations == nil {
+		return false
+	}
+	if existingOwners, ok := annotations[s.ownerAnnotationKey]; ok {
+		for _, ownerReference := range strings.Split(existingOwners, ",") {
+			if ownerReference == ownerAnnotation {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+func (s *Synchronizer[T, P]) addOwnerAnnotation(obj client.Object, owner client.Object) {
+	ownerAnnotation := s.makeOwnerAnnotation(owner)
+	annotations := obj.GetAnnotations()
+	if annotations == nil {
+		annotations = make(map[string]string)
+	}
+	var ownerReferences []string
+	if existingOwners, ok := annotations[s.ownerAnnotationKey]; ok {
+		ownerReferences = strings.Split(existingOwners, ",")
+	}
+	ownerReferences = append(ownerReferences, ownerAnnotation)
+	annotations[s.ownerAnnotationKey] = strings.Join(ownerReferences, ",")
+	obj.SetAnnotations(annotations)
+}
+
+func (s *Synchronizer[T, P]) removeOwnerAnnotation(obj client.Object, owner client.Object) {
+	ownerAnnotation := s.makeOwnerAnnotation(owner)
+	ownerReferences := s.GetOwnerAnnotations(obj)
+	newOwnerReferences := make([]string, len(ownerReferences))
+	for _, ownerReference := range ownerReferences {
+		if ownerReference == ownerAnnotation {
+			continue
+		}
+		newOwnerReferences = append(newOwnerReferences, ownerReference)
+	}
+	annotations := obj.GetAnnotations()
+	if annotations == nil {
+		annotations = make(map[string]string)
+	}
+	if len(newOwnerReferences) == 0 {
+		delete(annotations, s.ownerAnnotationKey)
+	} else {
+		annotations[s.ownerAnnotationKey] = strings.Join(newOwnerReferences, ",")
+	}
+	obj.SetAnnotations(annotations)
+}
+
+func (s *Synchronizer[T, P]) GetOwnerAnnotations(obj client.Object) []string {
+	annotations := obj.GetAnnotations()
+	if ownerAnnotations, ok := annotations[s.ownerAnnotationKey]; ok {
+		if ownerAnnotations == "" {
+			return []string{}
+		}
+		return strings.Split(ownerAnnotations, ",")
+	}
+	return []string{}
+}
 
 // SetupWithManager sets up the controller with the Manager.
 func (s *Synchronizer[T, P]) SetupWithManager(mgr ctrl.Manager) error {
@@ -263,20 +319,20 @@ func (s *Synchronizer[T, P]) SetupWithManager(mgr ctrl.Manager) error {
 
 	for _, t := range s.reconciler.AdditionalTypes() {
 		builder = builder.Watches(t, handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, object client.Object) []reconcile.Request {
-			if value, ok := object.GetAnnotations()[s.ownerAnnotationKey]; ok {
-				name, err := parseNamespacedName(value)
-				if err != nil {
-					mgr.GetLogger().Error(err, "unable to parse owner")
-					return nil
-				}
-
-				return []reconcile.Request{
-					{
+			requests := make([]reconcile.Request, 0)
+			if ownerReferences, ok := object.GetAnnotations()[s.ownerAnnotationKey]; ok {
+				for _, ownerReference := range strings.Split(ownerReferences, ",") {
+					name, err := parseNamespacedName(ownerReference)
+					if err != nil {
+						mgr.GetLogger().Error(err, "unable to parse owner")
+						continue
+					}
+					requests = append(requests, reconcile.Request{
 						NamespacedName: name,
-					},
+					})
 				}
 			}
-			return nil
+			return requests
 		}))
 	}
 	return builder.
@@ -286,7 +342,6 @@ func (s *Synchronizer[T, P]) SetupWithManager(mgr ctrl.Manager) error {
 func (s *Synchronizer[T, P]) DetectUnreferenced(ctx context.Context, owner T, actions []action.Action) ([]action.Action, error) {
 	// List all resources of owned or additional types
 	// Filter unrelated resources (owner annotation / owner reference)
-	annotationValue := client.ObjectKeyFromObject(owner).String()
 	allResources := make([]client.Object, 0)
 	for _, t := range s.relevantListTypes {
 		list := reflect.New(t).Interface().(client.ObjectList)
@@ -296,11 +351,8 @@ func (s *Synchronizer[T, P]) DetectUnreferenced(ctx context.Context, owner T, ac
 		}
 		err = meta.EachListItem(list, func(obj runtime.Object) error {
 			if cObj, ok := obj.(client.Object); ok {
-				annotations := cObj.GetAnnotations()
-				if v, ok := annotations[s.ownerAnnotationKey]; ok {
-					if v == annotationValue {
-						allResources = append(allResources, cObj)
-					}
+				if s.HasOwnerAnnotation(cObj, owner) {
+					allResources = append(allResources, cObj)
 				}
 			}
 			return nil
@@ -328,9 +380,15 @@ func (s *Synchronizer[T, P]) DetectUnreferenced(ctx context.Context, owner T, ac
 			unreferenced = append(unreferenced, existing)
 		}
 	}
-	// Add DeleteIfExists action for remainder
+
+	// Remove owner annotation on shared resources, delete if last owner
 	for _, existing := range unreferenced {
-		actions = append(actions, action.DeleteIfExists(existing, owner, func(obj client.Object, _ *runtime.Scheme) []meta_v1.Condition { return nil }, s.recorder))
+		s.removeOwnerAnnotation(existing, owner)
+		if len(s.GetOwnerAnnotations(existing)) > 0 {
+			actions = append(actions, action.Update(existing, owner, func(obj client.Object, _ *runtime.Scheme) []meta_v1.Condition { return nil }, s.recorder))
+		} else {
+			actions = append(actions, action.DeleteIfExists(existing, owner, func(obj client.Object, _ *runtime.Scheme) []meta_v1.Condition { return nil }, s.recorder))
+		}
 	}
 
 	return actions, nil
